Implement JPEG w/o resizing

+++ src/cli.rs
@@ -27,6 +27,6 @@ pub struct Args {
     /// JPEG quality, on a scale from 1 (worst) to 100 (best);
     /// ignored in case of PNGs
     #[arg(short, long, default_value_t = QUALITY,
-    value_parser = clap::value_parser!(u8).range(1..=100))]
+    value_parser = clap::value_parser!(i32).range(1..=100))]
     pub quality: i32,
 }
